Playerbot: Use references for targets
All functions taking a target parameter did so by using a nullable pointer.
This has led to a lot of defensive programming, because it has become unclear whether a target was passed in.
Using references makes it clear that a target was passed in.

This avoids a lot of error conditions (checking for null), and dead code in gameplay logic (what to do when there's no target).
It also avoids unnecessary searches for the target by GUID, when the bot already has that information.
Some instances of missing checks for null are now proven to be safe as well.